requirements.txt
boto3
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
